Rename `DatabaseEncryption` class to `DatabaseEncryptor`
