+++ app/src/androidTest/java/org/simple/clinic/di/TestAppComponent.kt
@@ -46,7 +46,7 @@ import org.simple.clinic.rules.ServerRegistrationAtFacilityRule
 import org.simple.clinic.security.pin.BruteForceProtectionAndroidTest
 import org.simple.clinic.signature.SignatureRepositoryAndroidTest
 import org.simple.clinic.storage.DaoWithUpsertAndroidTest
-import org.simple.clinic.storage.DatabaseEncryptionTest
+import org.simple.clinic.storage.DatabaseEncryptorTest
 import org.simple.clinic.storage.DeleteSyncGroupDatabaseAndroidTest
 import org.simple.clinic.storage.PurgeDatabaseAndroidTest
 import org.simple.clinic.storage.migrations.BaseDatabaseMigrationTest
@@ -161,5 +161,5 @@ interface TestAppComponent {
   fun inject(target: QuestionnaireSyncIntegrationTest)
   fun inject(target: QuestionnaireResponseRepositoryAndroidTest)
   fun inject(target: QuestionnaireResponseSyncIntegrationTest)
-  fun inject(target: DatabaseEncryptionTest)
+  fun inject(target: DatabaseEncryptorTest)
 }
