+++ app/src/androidTest/java/org/simple/clinic/storage/DatabaseEncryptorTest.kt
@@ -5,15 +5,14 @@ import androidx.room.Room
 import com.google.common.truth.Truth.assertThat
 import com.squareup.moshi.Moshi
 import net.sqlcipher.database.SupportFactory
-import org.junit.After
 import org.junit.Before
 import org.junit.Test
 import org.simple.clinic.AppDatabase
 import org.simple.clinic.TestClinicApp
 import org.simple.clinic.questionnaire.component.BaseComponentData
-import org.simple.clinic.storage.DatabaseEncryption.State.DOES_NOT_EXIST
-import org.simple.clinic.storage.DatabaseEncryption.State.ENCRYPTED
-import org.simple.clinic.storage.DatabaseEncryption.State.UNENCRYPTED
+import org.simple.clinic.storage.DatabaseEncryptor.State.DOES_NOT_EXIST
+import org.simple.clinic.storage.DatabaseEncryptor.State.ENCRYPTED
+import org.simple.clinic.storage.DatabaseEncryptor.State.UNENCRYPTED
 import org.simple.clinic.user.User
 import org.simple.clinic.user.UserStatus
 import org.simple.sharedTestCode.TestData
@@ -21,13 +20,13 @@ import java.time.Instant
 import java.util.UUID
 import javax.inject.Inject
 
-class DatabaseEncryptionTest {
+class DatabaseEncryptorTest {
 
   @Inject
   lateinit var appContext: Application
 
   @Inject
-  lateinit var databaseEncryption: DatabaseEncryption
+  lateinit var databaseEncryptor: DatabaseEncryptor
 
   @Inject
   lateinit var moshi: Moshi
@@ -45,9 +44,9 @@ class DatabaseEncryptionTest {
   @Test
   fun encrypting_existing_database_should_work_correctly() {
     // given
-    val passphrase = databaseEncryption.passphrase
+    val passphrase = databaseEncryptor.passphrase
 
-    assertThat(databaseEncryption.databaseState(DB_NAME)).isEqualTo(DOES_NOT_EXIST)
+    assertThat(databaseEncryptor.databaseState(DB_NAME)).isEqualTo(DOES_NOT_EXIST)
 
     val unencryptedDatabase = Room.databaseBuilder(appContext, AppDatabase::class.java, DB_NAME)
         .allowMainThreadQueries()
@@ -72,13 +71,13 @@ class DatabaseEncryptionTest {
 
     unencryptedDatabase.close()
 
-    assertThat(databaseEncryption.databaseState(DB_NAME)).isEqualTo(UNENCRYPTED)
+    assertThat(databaseEncryptor.databaseState(DB_NAME)).isEqualTo(UNENCRYPTED)
 
     // when
-    databaseEncryption.execute(databaseName = DB_NAME)
+    databaseEncryptor.execute(databaseName = DB_NAME)
 
     // then
-    assertThat(databaseEncryption.databaseState(DB_NAME)).isEqualTo(ENCRYPTED)
+    assertThat(databaseEncryptor.databaseState(DB_NAME)).isEqualTo(ENCRYPTED)
 
     val encryptedDatabase = Room.databaseBuilder(appContext, AppDatabase::class.java, DB_NAME)
         .allowMainThreadQueries()
